Refactor python build script Fixed wrongly patch dart tool/utils.py for python >= 3.12 Add running blutter without libflutter by specify dart version

+++ dartvm_fetch_build.py
@@ -31,8 +31,23 @@ def load_source(modname, filename):
     return module
 """
 
-def checkout_dart(ver: str, snapshot_hash: str = None):
-    clonedir = os.path.join(SDK_DIR, 'v'+ver)
+class DartLibInfo:
+    def __init__(self, version: str, os_name: str, arch: str, has_compressed_ptrs: bool = None, snapshot_hash: str = None):
+        self.version = version
+        self.os_name = os_name
+        self.arch = arch
+        self.snapshot_hash = snapshot_hash
+        if has_compressed_ptrs is None:
+            # use same as flutter default configuration
+            # TODO: old Dart version has no pointer compression
+            self.has_compressed_ptrs = os_name != 'ios'
+        else:
+            self.has_compressed_ptrs = has_compressed_ptrs
+        self.lib_name = f'dartvm{version}_{os_name}_{arch}'
+
+
+def checkout_dart(info: DartLibInfo):
+    clonedir = os.path.join(SDK_DIR, 'v'+info.version)
 
     # if no version file,assume previous clone is failed. delete the whole directory and try again.
     version_file = os.path.join(clonedir, 'runtime', 'vm', 'version.cc')
@@ -46,7 +61,7 @@ def checkout_dart(ver: str, snapshot_hash: str = None):
     # clone Dart source code
     if not os.path.exists(clonedir):
         # minimum clone repository at the target branch
-        subprocess.run([GIT_CMD, '-c', 'advice.detachedHead=false', 'clone', '-b', ver, '--depth', '1', '--filter=blob:none', '--sparse', '--progress', DART_GIT_URL, clonedir], check=True)
+        subprocess.run([GIT_CMD, '-c', 'advice.detachedHead=false', 'clone', '-b', info.version, '--depth', '1', '--filter=blob:none', '--sparse', '--progress', DART_GIT_URL, clonedir], check=True)
         # checkout only needed sources (runtime and tools)
         # since Dart 3.3 "third_party/double-conversion" is moved to outside of "runtime" directory
         subprocess.run([GIT_CMD, 'sparse-checkout', 'set', 'runtime', 'tools', 'third_party/double-conversion'], cwd=clonedir, check=True)
@@ -59,7 +74,7 @@ def checkout_dart(ver: str, snapshot_hash: str = None):
                     # should ".git" directory be removed?
                     pass
 
-        if snapshot_hash is None:
+        if info.snapshot_hash is None:
             # if running with Python 3.12, tools/utils.py should be patched to replace imp module with importlib
             # due to its remotion as stated in: https://docs.python.org/3.12/whatsnew/3.12.html#imp
             if sys.version_info[:2] >= (3, 12):
@@ -67,32 +82,32 @@ def checkout_dart(ver: str, snapshot_hash: str = None):
                 if os.path.exists(utils_path):
                     with open(utils_path, "r+") as f:
                         content = f.read()
-                        if "import imp\n" in content:
-                            content = content.replace("import imp\n", imp_replace_snippet).replace("imp.load_source", "load_source")
-                        # replace invalid escape sequences strings with raw strings to avoid SyntaxWarning
-                        # in future Python versions this warning will raise an error instead of warning
-                        # as stated in: https://docs.python.org/3/whatsnew/3.12.html#other-language-changes
-                        content = content.replace(" ' awk ", " r' awk ").replace("match_against('", "match_against(r'").replace("re.search('", "re.search(r'")
-                        f.seek(0)
-                        f.truncate()
-                        f.write(content)
+                        # the python 3.12 warnings are fixed in https://github.com/dart-lang/sdk/commit/a100968232c7492ab0de26897ff019e5784d4d38
+                        if r"match_against('^MAJOR (\d+)$', content)" in content:
+                            # old Dart version
+                            # replace invalid escape sequences strings with raw strings to avoid SyntaxWarning
+                            # in future Python versions this warning will raise an error instead of warning
+                            # as stated in: https://docs.python.org/3/whatsnew/3.12.html#other-language-changes
+                            content = content.replace(" ' awk ", " r' awk ").replace("match_against('", "match_against(r'").replace("re.search('", "re.search(r'")
+                            if "import imp\n" in content:
+                                content = content.replace("import imp\n", imp_replace_snippet).replace("imp.load_source", "load_source")
+                            f.seek(0)
+                            f.truncate()
+                            f.write(content)
             # make version
             subprocess.run([sys.executable, 'tools/make_version.py', '--output', 'runtime/vm/version.cc', '--input', 'runtime/vm/version_in.cc'], cwd=clonedir, check=True)
         else:
-            subprocess.run([sys.executable, MAKE_VERSION_FILE, clonedir, snapshot_hash], check=True)
+            subprocess.run([sys.executable, MAKE_VERSION_FILE, clonedir, info.snapshot_hash], check=True)
     
     return clonedir
 
-def get_dartlib_name(ver: str, arch: str, os_name: str):
-    return f'dartvm{ver}_{os_name}_{arch}'
-
-def cmake_dart(ver: str, arch: str, os_name: str, has_compressed_ptrs: bool, target_dir: str):
+def cmake_dart(info: DartLibInfo, target_dir: str):
     # On windows, need developer command prompt for x64 (can check with "cl" command)
     # create dartsdk/vx.y.z/CMakefile.list
     with open(CMAKE_TEMPLATE_FILE, 'r') as f:
         code = f.read()
     with open(os.path.join(target_dir, 'CMakeLists.txt'), 'w') as f:
-        f.write(code.replace('VERSION_PLACE_HOLDER', ver))
+        f.write(code.replace('VERSION_PLACE_HOLDER', info.version))
 
     # create dartsdk/vx.y.z/Config.cmake.in
     with open(os.path.join(target_dir, 'Config.cmake.in'), 'w') as f:
@@ -106,22 +121,23 @@ def cmake_dart(ver: str, arch: str, os_name: str, has_compressed_ptrs: bool, tar
     #   add -DTARGET_OS=ios for analyzing ios App
     # Note: pointer compression feature is set from Flutter and no one change it when building app.
     #       so only one build of Dart runtime is enough
-    builddir = os.path.join(BUILD_DIR, get_dartlib_name(ver, arch, os_name))
-    subprocess.run([CMAKE_CMD, '-GNinja', '-B', builddir, f'-DTARGET_OS={os_name}', f'-DTARGET_ARCH={arch}', 
-        f'-DCOMPRESSED_PTRS={1 if has_compressed_ptrs else 0}', '-DCMAKE_BUILD_TYPE=Release', '--log-level=NOTICE'], 
+    builddir = os.path.join(BUILD_DIR, info.lib_name)
+    subprocess.run([CMAKE_CMD, '-GNinja', '-B', builddir, f'-DTARGET_OS={info.os_name}', f'-DTARGET_ARCH={info.arch}', 
+        f'-DCOMPRESSED_PTRS={1 if info.has_compressed_ptrs else 0}', '-DCMAKE_BUILD_TYPE=Release', '--log-level=NOTICE'], 
         cwd=target_dir, check=True)
     
     # build and install dart vm library to packages directory
     subprocess.run([NINJA_CMD], cwd=builddir, check=True)
     subprocess.run([CMAKE_CMD, '--install', '.'], cwd=builddir, check=True)
 
-def fetch_and_build(ver: str, arch: str, os_name: str, has_compressed_ptrs: bool, snapshot_hash: str = None):
-    outdir = checkout_dart(ver, snapshot_hash)
-    cmake_dart(ver, arch, os_name, has_compressed_ptrs, outdir)
+def fetch_and_build(info: DartLibInfo):
+    outdir = checkout_dart(info)
+    cmake_dart(info, outdir)
 
 if __name__ == "__main__":
     ver = sys.argv[1]
     os_name = 'android' if len(sys.argv) < 3 else sys.argv[2]
     arch = 'arm64' if len(sys.argv) < 4 else sys.argv[3]
     snapshot_hash = None if len(sys.argv) < 5 else sys.argv[4]
-    fetch_and_build(ver, arch, os_name, True, snapshot_hash)
+    info = DartLibInfo(ver, os_name, arch, snapshot_hash=snapshot_hash)
+    fetch_and_build(info)
